refactor: controller 역할에 따라 분리
